Refactor:#76 - standardize timer icon naming, location, and format
- Relocate from public/assets/icons/timerIcon.svg to src/assets/icons/timer.svg
- Migrate from next/Image to SVGR import
- Apply consistent SVGR props
