+++ src/components/insight/InsightCard.tsx
@@ -1,5 +1,3 @@
-import Image from 'next/image';
-
 import { cn } from '@/lib/utils';
 
 interface InsightCardProps {
@@ -7,6 +5,8 @@ interface InsightCardProps {
   items?: React.ReactNode[];
   className?: string;
 }
+
+import TimerIcon from '@/assets/icons/timer.svg';
 
 const InsightCard = ({ variant, items = [], className }: InsightCardProps) => {
   const baseClasses =
@@ -16,7 +16,8 @@ const InsightCard = ({ variant, items = [], className }: InsightCardProps) => {
     return (
       <div className={cn(baseClasses, 'items-center justify-center text-center', className)}>
         <div className="text-body-m-16 text-text-03 flex flex-col items-center gap-8">
-          <Image src="/assets/icons/timerIcon.svg" width={32} height={32} alt="타이머 아이콘" />
+          <TimerIcon className="text-text-inactive" width={32} height={32} fill="currentColor" />
+
           <p>
             이번 주 작업 기록이 없어 <br className="md:hidden" />
             인사이트를 제공할 수 없어요 :(
