+++ src/components/timer/SelectTodoModal.tsx
@@ -2,7 +2,7 @@
 
 import { useEffect, useState } from 'react';
 
-import TimerIcon from '@/../public/assets/icons/timerIcon.svg';
+import TimerIcon from '@/assets/icons/timer.svg';
 import { IconButton } from '@/components/ui/IconButton';
 import Modal from '@/components/ui/Modal';
 import { GoalSummary } from '@/interfaces/goal';
@@ -32,7 +32,7 @@ export default function SelectTodoModal({
     <Modal isOpen onClose={onClose} size="timer">
       <div className="mb-40 flex items-center justify-between">
         <div className="flex items-center gap-12">
-          <TimerIcon className="text-Gray_01 h-24 w-24" />
+          <TimerIcon className="text-snackbar" width={24} height={24} fill="currentColor" />
           <h2 className="text-display-24 text-text-01">어떤 작업을 시작해볼까요?</h2>
         </div>
         <IconButton variant="close" aria-label="닫기" onClick={onClose} />
