+++ src/components/timer/TimerHeader.tsx
@@ -1,4 +1,4 @@
-import TimerIcon from '@/../public/assets/icons/timerIcon.svg';
+import TimerIcon from '@/assets/icons/timer.svg';
 import { IconButton } from '@/components/ui/IconButton';
 
 interface TimerHeaderProps {
@@ -11,7 +11,7 @@ export default function TimerHeader({ onBack, onClose }: TimerHeaderProps) {
     <div className="mb-40 flex items-center justify-between">
       <div className="flex items-center gap-12">
         <IconButton variant="back" aria-label="뒤로가기" onClick={onBack} />
-        <TimerIcon className="text-Gray_01 h-24 w-24" />
+        <TimerIcon className="text-snackbar" width={24} height={24} fill="currentColor" />
         <h2 className="text-display-24">할 일 타이머</h2>
       </div>
       <IconButton variant="close" aria-label="닫기" onClick={onClose} />
