+++ src/stories/ui/Popover.stories.tsx
@@ -1,8 +1,8 @@
 import { useRef, useState } from 'react';
 
-import Image from 'next/image';
 import type { Meta, StoryObj } from '@storybook/nextjs-vite';
 
+import TimerIcon from '@/assets/icons/timer.svg';
 import LegendSection from '@/components/heatmaps/LegendSection';
 import Popover from '@/components/ui/Popover';
 import { MONTHLY_LEGEND, WEEKLY_LEGEND } from '@/constants/heatmap';
@@ -91,9 +91,7 @@ export const HeatmapDefault: Story = {
           title="기본 히트맵 팝오버"
           variant="heatmap"
           onClose={() => setIsOpen(false)}
-          icon={
-            <Image src="/assets/icons/timerIcon.svg" alt="타이머 아이콘" width={24} height={24} />
-          }
+          icon={<TimerIcon className="text-snackbar" width={24} height={24} fill="currentColor" />}
         >
           <div className="flex flex-row gap-76">
             <LegendSection title="[주간]" data={WEEKLY_LEGEND} />
